+++ gates/generate-showcase.mjs
@@ -135,6 +135,7 @@ function showcase() {
         <link rel="stylesheet" href="../css/retro-register.css" />
         <link rel="stylesheet" href="../css/synthwave-register.css" />
         <link rel="stylesheet" href="../css/phantom-register.css" />
+        <link rel="stylesheet" href="../css/terminal-register.css" />
         <style>
 ${STYLE}        </style>
     </head>
@@ -234,6 +235,7 @@ function fixture(theme) {
         <link rel="stylesheet" href="../../css/retro-register.css" />
         <link rel="stylesheet" href="../../css/synthwave-register.css" />
         <link rel="stylesheet" href="../../css/phantom-register.css" />
+        <link rel="stylesheet" href="../../css/terminal-register.css" />
         <style>
 ${STYLE}        </style>
     </head>